requirements.txt
numpy==1.23.3
Pillow==9.2.0
tflite-runtime==2.10.0
